requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]>=3.1.0
Pillow>=10.0.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
django-filter==23.5
